refactor: Add head prop to DataGrid component
This commit adds a new prop called "head" to the DataGrid component in order to allow custom rendering of the table header. The "head" prop accepts a React node and can be used to display additional content or customize the appearance of the table header. This enhancement provides more flexibility and customization options for the DataGrid component.