feat: Include Avatar component in @mui/joy exports
This commit adds the Avatar component to the exports of the "@mui/joy" package. The Avatar component is used to display user profile pictures or icons. By including the Avatar component in the exports, it provides users with a convenient way to use this functionality in their applications. This enhancement improves the usability and flexibility of the "@mui/joy" package.